[ANTHA-1972] Initial implementation attempt. Please see the JIRA ticket for discussion of the design approach.
